1. Add remove option to Menu 2. Add getIndex method to Menu 3. Get assignment ops (lvalue & rvalue refs) to Rectangle and Square 4. Some random changes

+++ C++/Vertex.h
@@ -4,7 +4,7 @@
 #include <stdexcept>
 #include <math.h>
 
-using _Vertex_t = int;
+using _Vertex_t = float;
 
 class Vertex {
 private:
